requirements.txt
streamlit
streamlit-option-menu
matplotlib
pymongo
python-dotenv
fastapi
uvicorn